+++ .idea/misc.xml
@@ -10,18 +10,17 @@
         <entry key="..\:/YULIM/Android-Yurim/app/src/main/res/drawable/btn_rectangle.xml" value="0.1" />
         <entry key="..\:/YULIM/Android-Yurim/app/src/main/res/drawable/ic_img_github.xml" value="0.38229166666666664" />
         <entry key="..\:/YULIM/Android-Yurim/app/src/main/res/drawable/sd_rectangle.xml" value="0.275" />
+        <entry key="..\:/YULIM/Android-Yurim/app/src/main/res/drawable/selector_checkbox.xml" value="0.3203125" />
         <entry key="..\:/YULIM/Android-Yurim/app/src/main/res/drawable/selector_list_btn.xml" value="0.1" />
         <entry key="..\:/YULIM/Android-Yurim/app/src/main/res/drawable/selector_tv.xml" value="0.10185185185185185" />
         <entry key="..\:/YULIM/Android-Yurim/app/src/main/res/drawable/tv_rectangle.xml" value="0.287962962962963" />
         <entry key="..\:/YULIM/Android-Yurim/app/src/main/res/drawable/tv_rectangle_pink.xml" value="0.18518518518518517" />
-        <entry key="..\:/YULIM/Android-Yurim/app/src/main/res/layout/activity_find_user.xml" value="0.1911231884057971" />
         <entry key="..\:/YULIM/Android-Yurim/app/src/main/res/layout/activity_home.xml" value="0.25" />
         <entry key="..\:/YULIM/Android-Yurim/app/src/main/res/layout/activity_sign_in.xml" value="0.25" />
         <entry key="..\:/YULIM/Android-Yurim/app/src/main/res/layout/activity_sign_up.xml" value="0.33" />
         <entry key="..\:/YULIM/Android-Yurim/app/src/main/res/layout/follower_list.xml" value="0.36614583333333334" />
-        <entry key="..\:/YULIM/Android-Yurim/app/src/main/res/layout/fragment_follower.xml" value="0.1" />
-        <entry key="..\:/YULIM/Android-Yurim/app/src/main/res/layout/fragment_home.xml" value="0.15353260869565216" />
-        <entry key="..\:/YULIM/Android-Yurim/app/src/main/res/layout/fragment_profile.xml" value="0.25" />
+        <entry key="..\:/YULIM/Android-Yurim/app/src/main/res/layout/fragment_follower.xml" value="0.25" />
+        <entry key="..\:/YULIM/Android-Yurim/app/src/main/res/layout/fragment_profile.xml" value="0.33" />
         <entry key="..\:/YULIM/Android-Yurim/app/src/main/res/layout/fragment_repository.xml" value="0.17314814814814813" />
         <entry key="..\:/YULIM/Android-Yurim/app/src/main/res/layout/repository_list.xml" value="0.25" />
       </map>
