+++ app/src/main/java/kr/co/softcampus/sopt_assignment1/Activity/SignUpActivity.kt
@@ -1,10 +1,14 @@
-package kr.co.softcampus.sopt_assignment1
+package kr.co.softcampus.sopt_assignment1.Activity
 
 import android.content.Intent
 import android.os.Bundle
 import android.util.Log
 import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
+import kr.co.softcampus.sopt_assignment1.Data.RequestSignupData
+import kr.co.softcampus.sopt_assignment1.Data.ResponseSignupData
+import kr.co.softcampus.sopt_assignment1.R
+import kr.co.softcampus.sopt_assignment1.ServiceCreator
 import kr.co.softcampus.sopt_assignment1.databinding.ActivitySignUpBinding
 import retrofit2.Call
 import retrofit2.Callback
