+++ app/src/main/java/kr/co/softcampus/sopt_assignment1/Adapter/HomeViewPagerAdapter.kt
@@ -1,4 +1,4 @@
-package kr.co.softcampus.sopt_assignment1
+package kr.co.softcampus.sopt_assignment1.Adapter
 
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentActivity
