+++ app/src/main/java/kr/co/softcampus/sopt_assignment1/Adapter/RepositoryAdapter.kt
@@ -1,8 +1,9 @@
-package kr.co.softcampus.sopt_assignment1
+package kr.co.softcampus.sopt_assignment1.Adapter
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
+import kr.co.softcampus.sopt_assignment1.Data.RepositoryData
 import kr.co.softcampus.sopt_assignment1.databinding.RepositoryListBinding
 
 class RepositoryAdapter : RecyclerView.Adapter<RepositoryAdapter.RepositoryViewHolder>() {
