+++ app/src/main/java/kr/co/softcampus/sopt_assignment1/Data/FollowerData.kt
@@ -1,4 +1,4 @@
-package kr.co.softcampus.sopt_assignment1
+package kr.co.softcampus.sopt_assignment1.Data
 
 data class FollowerData(
     val name: String,
