+++ app/src/main/java/kr/co/softcampus/sopt_assignment1/Data/RepositoryData.kt
@@ -1,4 +1,4 @@
-package kr.co.softcampus.sopt_assignment1
+package kr.co.softcampus.sopt_assignment1.Data
 
 data class RepositoryData(
     val title: String,
