+++ app/src/main/java/kr/co/softcampus/sopt_assignment1/Data/RequestSignupData.kt
@@ -1,4 +1,4 @@
-package kr.co.softcampus.sopt_assignment1
+package kr.co.softcampus.sopt_assignment1.Data
 
 import com.google.gson.annotations.SerializedName
