+++ app/src/main/java/kr/co/softcampus/sopt_assignment1/Data/ResponseGetUserByEmailData.kt
@@ -1,8 +1,8 @@
-package kr.co.softcampus.sopt_assignment1
+package kr.co.softcampus.sopt_assignment1.Data
 
-data class ResponseSigninData( //생성자 소괄호
+data class ResponseGetUserByEmailData(
     val status: Int,
-    val success: Boolean,
+    val sucess: Boolean,
     val message: String,
     val data: Data
 ) {
