+++ app/src/main/java/kr/co/softcampus/sopt_assignment1/Data/ResponseSigninData.kt
@@ -1,6 +1,6 @@
-package kr.co.softcampus.sopt_assignment1
+package kr.co.softcampus.sopt_assignment1.Data
 
-data class ResponseSignupData(
+data class ResponseSigninData( //생성자 소괄호
     val status: Int,
     val success: Boolean,
     val message: String,
