+++ app/src/main/java/kr/co/softcampus/sopt_assignment1/Fragment/HomeFragment.kt
@@ -1,4 +1,4 @@
-package kr.co.softcampus.sopt_assignment1
+package kr.co.softcampus.sopt_assignment1.Fragment
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,6 +7,8 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import com.bumptech.glide.Glide
 import com.google.android.material.tabs.TabLayoutMediator
+import kr.co.softcampus.sopt_assignment1.Adapter.TabViewPagerAdapter
+import kr.co.softcampus.sopt_assignment1.R
 import kr.co.softcampus.sopt_assignment1.databinding.FragmentHomeBinding
 
 class HomeFragment : Fragment() {
