+++ app/src/main/java/kr/co/softcampus/sopt_assignment1/Fragment/TabFollowerFragment.kt
@@ -1,17 +1,19 @@
-package kr.co.softcampus.sopt_assignment1
+package kr.co.softcampus.sopt_assignment1.Fragment
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
+import kr.co.softcampus.sopt_assignment1.R
 
-class TabFollowingFragment : Fragment() {
+class TabFollowerFragment : Fragment() {
+
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_tab_following, container, false)
+        return inflater.inflate(R.layout.fragment_tab_follower, container, false)
     }
+
 }
