+++ app/src/main/java/kr/co/softcampus/sopt_assignment1/Fragment/TabFollowingFragment.kt
@@ -1,18 +1,18 @@
-package kr.co.softcampus.sopt_assignment1
+package kr.co.softcampus.sopt_assignment1.Fragment
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
+import kr.co.softcampus.sopt_assignment1.R
 
-class TabFollowerFragment : Fragment() {
-
+class TabFollowingFragment : Fragment() {
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        return inflater.inflate(R.layout.fragment_tab_follower, container, false)
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_tab_following, container, false)
     }
-
 }
