+++ app/src/main/java/kr/co/softcampus/sopt_assignment1/GetUserByEmailService.kt
@@ -1,5 +1,6 @@
 package kr.co.softcampus.sopt_assignment1
 
+import kr.co.softcampus.sopt_assignment1.Data.ResponseGetUserByEmailData
 import retrofit2.Call
 import retrofit2.http.GET
 import retrofit2.http.Headers
