+++ app/src/main/java/kr/co/softcampus/sopt_assignment1/SignupService.kt
@@ -1,5 +1,7 @@
 package kr.co.softcampus.sopt_assignment1
 
+import kr.co.softcampus.sopt_assignment1.Data.RequestSignupData
+import kr.co.softcampus.sopt_assignment1.Data.ResponseSignupData
 import retrofit2.Call
 import retrofit2.http.Body
 import retrofit2.http.Headers
