+++ app/src/main/res/layout/fragment_image.xml
@@ -5,7 +5,7 @@
     android:id="@+id/frameLayout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ImageFragment">
+    tools:context=".Fragment.ImageFragment">
 
     <ImageView
         android:id="@+id/iv_image"
