+++ app/src/main/res/layout/fragment_tab_following.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#DFDFDF"
-    tools:context=".TabFollowingFragment">
+    tools:context=".Fragment.TabFollowingFragment">
 
     <TextView
         android:layout_width="wrap_content"
